🚚 共通UIコンポーネントの整理
