+++ packages/ui/src/FuelEfficiencyChart/fuelChartTypes.ts
