+++ packages/ui/src/FuelEfficiencyChart/fuelChartUtils.ts
