--- packages/ui/src/code.tsx
@@ -1,11 +1,0 @@
-import { type JSX } from 'react'
-
-export function Code({
-  children,
-  className,
-}: {
-  children: React.ReactNode
-  className?: string
-}): JSX.Element {
-  return <code className={className}>{children}</code>
-}
